requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers>=4.7.0
djangorestframework-simplejwt>=5.5.0
python-decouple==3.8
Pillow==10.4.0
django-allauth==0.63.6
gunicorn==21.2.0
whitenoise==6.6.0
daphne==4.1.2
python-dotenv==1.0.0